refactor player_widget into a package with multiple files, fix labeling closest
