+++ src/jabs/ui/__init__.py
@@ -1,14 +1,3 @@
-from .about_dialog import AboutDialog
-from .central_widget import CentralWidget
-from .classification_thread import ClassifyThread
-from .frame_labels_widget import FrameLabelsWidget
-from .global_inference_widget import GlobalInferenceWidget
-from .identity_combo_box import IdentityComboBox
-from .k_fold_slider_widget import KFoldSliderWidget
-from .label_count_widget import FrameLabelCountWidget
 from .main_window import MainWindow
-from .manual_label_widget import ManualLabelWidget
-from .player_widget import PlayerWidget
-from .prediction_vis_widget import PredictionVisWidget
-from .timeline_label_widget import TimelineLabelWidget
-from .training_thread import TrainingThread
+
+__all__ = ["MainWindow"]
