+++ src/jabs/ui/player_widget/__init__.py
@@ -1,0 +1,5 @@
+from .player_widget import PlayerWidget
+
+__all__ = [
+    "PlayerWidget",
+]
