+++ src/jabs/ui/player_widget/player_widget.py
@@ -1,248 +1,10 @@
-import time
 from pathlib import Path
 
-import numpy as np
 from PySide6 import QtCore, QtGui, QtWidgets
 from PySide6.QtGui import QPaintEvent
 
-from jabs.feature_extraction.social_features.social_distance import ClosestIdentityInfo
-from jabs.pose_estimation import PoseEstimationV3, PoseEstimation
-from jabs.video_stream import (VideoReader, label_identity, label_all_identities,
-                               draw_track, overlay_pose, overlay_landmarks, overlay_segmentation)
-
-_CLOSEST_LABEL_COLOR = (255, 0, 0)
-_CLOSEST_FOV_LABEL_COLOR = (0, 255, 0)
-
-
-def _get_closest_animal_id(ref_id, frame_index, pose_est, half_fov_deg=None):
-
-    idx = PoseEstimationV3.KeypointIndex
-    closest_id = None
-    closest_dist = None
-    ref_shape = pose_est.get_identity_convex_hulls(ref_id)[frame_index]
-    if ref_shape is not None:
-        for curr_id in pose_est.identities:
-            if curr_id != ref_id:
-                other_shape = pose_est.get_identity_convex_hulls(curr_id)[frame_index]
-
-                if other_shape is not None:
-
-                    curr_dist = ref_shape.distance(other_shape)
-                    if half_fov_deg is None or half_fov_deg >= 180:
-
-                        # we can ignore FoV angle and just worry about distance
-                        if closest_dist is None or curr_dist < closest_dist:
-                            closest_id = curr_id
-                            closest_dist = curr_dist
-                    else:
-                        # we need to account for FoV angle
-                        points, mask = pose_est.get_points(frame_index, ref_id)
-
-                        # we need nose and base neck to figure out view angle
-                        if mask[idx.NOSE] == 1 and mask[idx.BASE_NECK] == 1:
-                            ref_base_neck_point = points[idx.BASE_NECK, :]
-                            ref_nose_point = points[idx.NOSE, :]
-                            other_centroid = np.array(other_shape.centroid)
-
-                            view_angle = ClosestIdentityInfo.compute_angle(
-                                ref_nose_point,
-                                ref_base_neck_point,
-                                other_centroid)
-
-                            # for FoV we want the range of view angle to be [180, -180)
-                            if view_angle > 180:
-                                view_angle -= 360
-
-                            if abs(view_angle) <= half_fov_deg:
-                                # other animal is in FoV
-                                if closest_dist is None or curr_dist < closest_dist:
-                                    closest_id = curr_id
-                                    closest_dist = curr_dist
-
-    return closest_id
-
-
-class _PlayerThread(QtCore.QThread):
-    """
-    thread used to grab frames (numpy arrays) from a video stream and convert
-    them to a QImage for display by the frame widget
-
-    handles timing to get correct playback speed
-    """
-
-    # signals used to update the UI components from the thread
-    newImage = QtCore.Signal(dict)
-    updatePosition = QtCore.Signal(dict)
-    endOfFile = QtCore.Signal()
-
-    def __init__(self, video_reader, pose_est, identity, show_track=False,
-                 overlay_pose_flag=False, identities=None, overlay_landmarks_flag=False,
-                 overlay_segmentation_flag=False):
-        super().__init__()
-        self._video_reader = video_reader
-        self._pose_est = pose_est
-        self._identity = identity
-        self._label_closest = False
-        self._show_track = show_track
-        self._overlay_pose = overlay_pose_flag
-        self._overlay_segmentation = overlay_segmentation_flag
-        self._overlay_landmarks = overlay_landmarks_flag
-        self._identities = identities if identities is not None else []
-        self._playing = False
-        self._lock = QtCore.QMutex()
-
-    def stop_playback(self):
-        """
-        tell run thread to stop playback
-        """
-        with QtCore.QMutexLocker(self._lock):
-            self._playing = False
-
-    def set_identity(self, identity):
-        """
-        set the active identity
-        :param identity: new selected identity
-        :return: None
-        """
-        self._identity = identity
-
-    def set_identities(self, identities):
-        self._identities = identities
-
-    def label_closest(self, new_val: bool):
-        self._label_closest = new_val
-
-    def set_show_track(self, new_val: bool):
-        self._show_track = new_val
-
-    def set_overlay_pose(self, new_val: bool):
-        self._overlay_pose = new_val
-
-    def set_overlay_segmentation(self, new_val: bool):
-        self._overlay_segmentation = new_val
-
-    def set_overlay_landmarks(self, new_val: bool):
-        self._overlay_landmarks = new_val
-
-    def _read_and_emit_frame(self):
-        frame = self._video_reader.load_next_frame()
-        image = self._prepare_image(frame)
-        self.newImage.emit({'image': image, 'source': self._video_reader.filename})
-        self.updatePosition.emit({'index': frame['index'], 'source': self._video_reader.filename})
-
-    def _prepare_image(self, frame: dict) -> QtGui.QImage | None:
-        if frame['data'] is None:
-            return None
-
-        if self._identity is not None:
-
-            if self._show_track:
-                draw_track(frame['data'], self._pose_est,
-                           self._identity, frame['index'])
-
-            if self._overlay_pose:
-                overlay_pose(
-                    frame['data'],
-                    *self._pose_est.get_points(frame['index'], self._identity)
-                )
-            if self._overlay_segmentation:
-                overlay_segmentation(
-                    frame['data'],
-                    self._pose_est,
-                    identity=self._identity,
-                    frameIndex=frame['index'],
-                    identities=self._identities
-                )
-            if self._overlay_landmarks:
-                overlay_landmarks(frame['data'], self._pose_est)
-
-            if self._label_closest:
-                closest_fov_id = _get_closest_animal_id(
-                    self._identity, frame['index'],
-                    self._pose_est, ClosestIdentityInfo._half_fov_deg)
-                if closest_fov_id is not None:
-                    label_identity(frame['data'], self._pose_est,
-                                   closest_fov_id, frame['index'],
-                                   color=_CLOSEST_FOV_LABEL_COLOR)
-
-                closest_id = _get_closest_animal_id(
-                    self._identity, frame['index'], self._pose_est)
-                if closest_id is not None and closest_id != closest_fov_id:
-                    label_identity(frame['data'], self._pose_est,
-                                   closest_id, frame['index'],
-                                   color=_CLOSEST_LABEL_COLOR)
-
-        # label all identities
-        label_all_identities(frame['data'],
-                             self._pose_est, self._identities,
-                             frame['index'], subject=self._identity)
-
-        # convert OpenCV image (numpy array) to QImage
-        image = QtGui.QImage(frame['data'], frame['data'].shape[1],
-                             frame['data'].shape[0],
-                             QtGui.QImage.Format_RGB888).rgbSwapped()
-
-        return image
-
-    def seek(self, position: int):
-        if not self._playing:
-            self._video_reader.seek(position)
-            self._read_and_emit_frame()
-
-    def load_new_video(self, video_reader: VideoReader, pose_est: PoseEstimation, identity: int, identities: list):
-        if not self._playing:
-            self._video_reader = video_reader
-            self._pose_est = pose_est
-            self._identity = identity
-            self._identities = identities
-
-    def run(self):
-        """
-        method to be run as a thread during playback
-        handles grabbing the next frame from the buffer, converting to a QImage,
-        and sending to the UI component for display.
-        """
-
-        # flag used to terminate loop after we've displayed the last frame
-        end_of_file = False
-        with QtCore.QMutexLocker(self._lock):
-            self._playing = True
-
-        next_timestamp = 0
-        start_time = 0
-
-        # iterate until we've been told to stop (user clicks pause button)
-        # or we reach end of file
-        while self._playing and not end_of_file:
-            now = time.perf_counter()
-            frame = self._video_reader.load_next_frame()
-            image = self._prepare_image(frame)
-
-            if image:
-                # don't update frame until we've shown the last one for the
-                # required duration
-                if start_time > 0:
-                    # sleep difference between next_timestamp and amount of
-                    # actual clock time since we started playback
-                    time.sleep(max(0, next_timestamp - (now - start_time)))
-                else:
-                    # first frame, save the start time
-                    start_time = now
-
-                # send the new frame and the frame index to the UI components
-                # unless playback was stopped while we were sleeping
-                if self._playing:
-                    self.newImage.emit({'image': image, 'source': self._video_reader.filename})
-                    self.updatePosition.emit({'index': frame['index'], 'source': self._video_reader.filename})
-
-                # update timestamp for when should the next frame be shown
-                next_timestamp += frame['duration']
-
-            else:
-                # if the video stream reached the end of file let the UI know
-                self.endOfFile.emit()
-                # and terminate the loop
-                end_of_file = True
+from .player_thread import PlayerThread
+from jabs.video_reader import VideoReader
 
 
 class _FrameWidget(QtWidgets.QLabel):
@@ -523,16 +285,23 @@ class PlayerWidget(QtWidgets.QWidget):
         assert self._video_stream is not None
         return self._video_stream.fps
 
-    def show_closest(self, new_val: bool | None):
-        if new_val is None:
+    def show_closest(self, enabled: bool | None = None):
+        """
+        change "show closest" state. Accepts a new boolean value, or toggles
+        """
+        if enabled is None:
             self._label_closest = not self._label_closest
         else:
-            self._label_closest = new_val
+            self._label_closest = enabled
 
         if self._player_thread:
             self._player_thread.label_closest(self._label_closest)
+            if not self._playing:
+                # if we are not playing, we need to update the current frame
+                # to show/hide the overlay
+                self._seek(self._position_slider.value())
 
-    def show_track(self, new_val: bool | None):
+    def show_track(self, new_val: bool | None = None):
         """
         change "show track" state. Accepts a new boolean value, or toggles
         current state if no value given.
@@ -544,6 +313,10 @@ class PlayerWidget(QtWidgets.QWidget):
 
         if self._player_thread:
             self._player_thread.set_show_track(self._show_track)
+            if not self._playing:
+                # if we are not playing, we need to update the current frame
+                # to show the overlay
+                self._seek(self._position_slider.value())
 
     def overlay_pose(self, new_val: bool | None):
         """
@@ -555,9 +328,6 @@ class PlayerWidget(QtWidgets.QWidget):
         else:
             self._overlay_pose = new_val
 
-        # don't do anything else if a video isn't loaded
-        if self._video_stream is None:
-            return
 
         if self._player_thread:
             self._player_thread.set_overlay_pose(self._overlay_pose)
@@ -633,7 +403,7 @@ class PlayerWidget(QtWidgets.QWidget):
         self._position_slider.setEnabled(True)
 
         if self._player_thread is None:
-            self._player_thread = _PlayerThread(
+            self._player_thread = PlayerThread(
                 self._video_stream, self._pose_est, self._active_identity,
                 self._show_track, self._overlay_pose, self._identities,
                 self._overlay_landmarks, self._overlay_segmentation)
