+++ src/jabs/video_reader/__init__.py
@@ -1,3 +1,13 @@
 from .frame_annotation import (label_identity, label_all_identities,
                                draw_track, overlay_pose, overlay_landmarks, overlay_segmentation)
 from .video_reader import VideoReader
+
+__all__ = [
+    'VideoReader',
+    'label_identity',
+    'label_all_identities',
+    'draw_track',
+    'overlay_pose',
+    'overlay_landmarks',
+    'overlay_segmentation',
+]
